requirements.txt
google-cloud-compute==1.14.1
functions-framework==3.5.0
